# docs/swagger.yaml
basePath: /api
definitions:
...
  models.XZXJUser:
    properties:
      experience:
        description: 个人经历
        type: string
      id:
        description: 学长学姐说集合中的_id
        type: string
      managedActivities:
        description: 主理活动
        type: string
      motto:
        description: 格言
        type: string
      picture:
        description: 学长学姐说用户照片
        type: string
      sections:
        description: 学长学姐说用户参与的分享计划
        items:
          type: string
        type: array
      userID:
        description: 参与学长学姐说的用户的真实ID，实际使用应该使用此ID
        type: string
    type: object
  svcerror.SvcErr:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  types.AddOrUpdateXZXJUserReq:
    properties:
      experience:
        description: 学长学姐说用户的个人经历
        type: string
      managedActivities:
        description: 学长学姐说用户的主理活动
        type: string
      motto:
        description: 学长学姐说用户的座右铭
        type: string
      picture:
        description: 学长学姐说用户的个人照片
        type: string
      realName:
        description: 学长学姐说用户的真实姓名
        type: string
      sections:
        description: 学长学姐说用户参与的分享计划
        items:
          type: string
        type: array
      userTags:
        description: 学长学姐说用户的标签
        items:
          type: string
        type: array
    type: object
  types.AddOrUpdateXZXJUserRes:
    properties:
      userID:
        description: 学长学姐说用户的用户ID
        type: string
    type: object
...
  types.DeleteXZXJUserRes:
    properties:
      deleted:
        type: boolean
    type: object
...
  types.QueryXZXJUserByUserIDRes:
    properties:
      xzxjUser:
        allOf:
        - $ref: '#/definitions/models.XZXJUser'
        description: 学长学姐说用户额外信息
    type: object
...
host: 127.0.0.1:8787
info:
  contact:
    email: enderbybear@foxmail.com
    name: Songyue Chen
  description: This is the backend for UniPingo application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: UniPingo Backend
  version: 0.0.1
paths:
...
  /xzxjdiscuss/add:
    post:
      description: 增加以为学长学姐说用户
      parameters:
      - description: JWT Token
        in: header
        name: x-access-token
        required: true
        type: string
      - description: 学长学姐说添加用户接口请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/types.AddOrUpdateXZXJUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 学长学姐说添加用户接口响应
          schema:
            $ref: '#/definitions/types.AddOrUpdateXZXJUserRes'
        "400":
          description: 请求格式问题
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
        "401":
          description: 用户认证失败
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
        "500":
          description: 服务器内部问题
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
      summary: 添加学长学姐说用户接口
      tags:
      - xzxj_discuss
  /xzxjdiscuss/delete:
    get:
      description: 删除学长学姐说用户
      parameters:
      - description: JWT Token
        in: header
        name: x-access-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除学长学姐说用户接口响应
          schema:
            $ref: '#/definitions/types.DeleteXZXJUserRes'
        "400":
          description: 请求格式问题
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
        "401":
          description: 用户认证失败
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
        "500":
          description: 服务器内部问题
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
      summary: 删除学长学姐说用户接口
      tags:
      - xzxj_discuss
  /xzxjdiscuss/get:
    get:
      description: 查询学长学姐说用户
      parameters:
      - description: JWT Token
        in: header
        name: x-access-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询学长学姐说用户接口响应
          schema:
            $ref: '#/definitions/types.QueryXZXJUserByUserIDRes'
        "400":
          description: 请求格式问题
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
        "401":
          description: 用户认证失败
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
        "500":
          description: 服务器内部问题
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
      summary: 查询学长学姐说用户接口
      tags:
      - xzxj_discuss
  /xzxjdiscuss/update:
    post:
      description: 增加以为学长学姐说用户
      parameters:
      - description: JWT Token
        in: header
        name: x-access-token
        required: true
        type: string
      - description: 学长学姐说添加用户接口请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/types.AddOrUpdateXZXJUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 学长学姐说添加用户接口响应
          schema:
            $ref: '#/definitions/types.AddOrUpdateXZXJUserRes'
        "400":
          description: 请求格式问题
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
        "401":
          description: 用户认证失败
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
        "500":
          description: 服务器内部问题
          schema:
            $ref: '#/definitions/svcerror.SvcErr'
      summary: 添加学长学姐说用户接口
      tags:
      - xzxj_discuss
schemes:
- https
swagger: "2.0"